requirements.txt
aiohttp==3.7.4.post0
alembic==1.6.2
amqp==5.0.6
astroid==2.5.7
async-timeout==3.0.1
attrs==21.2.0
autopep8==1.5.7
billiard==3.6.4.0
blinker==1.4
cachetools==4.2.2
celery==5.1.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
colorama==0.4.4
dateparser==1.0.0
Flask==2.0.0
Flask-Caching==1.10.1
Flask-Mail==0.9.1
flask-marshmallow==0.14.0
Flask-Migrate==3.0.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
google-api-core==1.28.0
google-auth==1.30.1
googleapis-common-protos==1.53.0
greenlet==1.1.0
gunicorn==20.1.0
idna==2.10
isort==5.8.0
itsdangerous==2.0.0
Jinja2==3.0.0
kombu==5.1.0
lazy-object-proxy==1.6.0
Mako==1.1.4
MarkupSafe==2.0.0
marshmallow==3.12.1
marshmallow-oneofschema==2.1.0
marshmallow-sqlalchemy==0.26.0
mccabe==0.6.1
multidict==5.1.0
opencensus==0.7.13
opencensus-context==0.1.2
opencensus-ext-azure==1.0.8
packaging==20.9
prompt-toolkit==3.0.18
protobuf==3.17.1
psutil==5.8.0
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pyodbc==4.0.30
pyparsing==2.4.7
python-binance==1.0.10
python-dateutil==2.8.1
python-decouple==3.4
python-dotenv==0.17.1
python-editor==1.0.4
pytz==2021.1
redis==3.5.3
regex==2021.4.4
requests==2.25.1
rsa==4.7.2
six==1.16.0
SQLAlchemy==1.4.15
toml==0.10.2
typing-extensions==3.10.0.0
tzlocal==2.1
ujson==4.0.2
urllib3==1.26.4
vine==5.0.0
wcwidth==0.2.5
websockets==9.0.2
Werkzeug==2.0.0
wrapt==1.12.1
yapf==0.31.0
yarl==1.6.3
